+++ tests/BufferTest.php
@@ -1,0 +1,30 @@
+<?php
+/**
+ * Copyright © EcomDev B.V. All rights reserved.
+ * See LICENSE for license details.
+ */
+
+declare(strict_types=1);
+
+namespace EcomDev\Bytes;
+
+use PHPUnit\Framework\TestCase;
+
+/**
+ * Class BufferTest
+ */
+class BufferTest extends TestCase
+{
+    /**
+     * @var BytesFactory
+     */
+    private $factory;
+
+    /**
+     *
+     */
+    protected function setUp(): void
+    {
+
+    }
+}